+++ CHANGELOG.md
@@ -22,6 +22,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Upgrade [pippo-jade] to Jade 1.1.4
 - Upgrade [pippo-pebble] to Pebble 2.2.1
 - Make `chunked` transfer-encoding optional, not the default
+- Make the text/plain content type engine handle returning reasonable types like String, CharSequence, char[], and byte[]
 
 #### Added
 - [#245]: Route groups